# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: Open Work Protocol (OWP) Pool API
  version: 0.1.0
  description: |
    Open Work Protocol (OWP) Pool is a capacity-aware task scheduler and worker protocol 
    that coordinates independent contributors using their own local AI coding tools to 
    produce PRs for shared repositories.
    
    ## Key Features
    - **BYO Seat Model**: Contributors use their own tools without sharing accounts
    - **Capacity-Aware Scheduling**: Points-based system with concurrency limits
    - **Maintainer Safety Controls**: Review budget throttling and area locks
    - **Resilient Task Management**: Leases with TTL and automatic requeue
    - **Advanced Features**: Reputation scoring, trust tiers, and analytics
  contact:
    name: OWP Pool API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8787
    description: Local development server
  - url: https://api.owp-pool.example.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Worker authentication token obtained from registration
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Administrative token for management operations
  schemas:
...
    WorkerStatus:
      type: string
      enum: [idle, working, paused]
      description: Current status of a worker
...
    RegisterWorkerRequest:
      type: object
      required: [name, skills, capacity_points, max_concurrent_tasks]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Human-readable worker name
        github_handle:
          type: string
          maxLength: 120
          nullable: true
          description: GitHub username for attribution
        skills:
          type: array
          items:
            type: string
          description: List of skills/technologies the worker can handle
        capacity_points:
          type: integer
          minimum: 1
          maximum: 100
          default: 5
          description: Daily capacity in points
        max_concurrent_tasks:
          type: integer
          minimum: 1
          maximum: 20
          default: 2
          description: Maximum number of concurrent task assignments
    RegisterWorkerResponse:
      type: object
      required: [worker_id, token]
      properties:
        worker_id:
          type: string
          description: Unique worker identifier
        token:
          type: string
          description: Bearer token for authentication
    HeartbeatRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/WorkerStatus'
          default: idle
        note:
          type: string
          maxLength: 500
          nullable: true
          description: Optional status message
    HeartbeatResponse:
      type: object
      required: [ok, server_time]
      properties:
        ok:
          type: boolean
          default: true
        server_time:
          type: string
          format: date-time
          description: Current server timestamp

    # Task Schemas
    TaskArtifact:
      type: object
      properties:
        pr_url:
          type: string
          format: uri
          nullable: true
          description: GitHub PR URL
        commit_sha:
          type: string
          nullable: true
          description: Git commit SHA
        patch_url:
          type: string
          format: uri
          nullable: true
          description: URL to patch file
        extra:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional artifact metadata
    LeaseView:
      type: object
      required: [task_id, repo, title, estimate_points, priority, required_skills, lease_expires_at]
      properties:
        task_id:
          type: string
          description: Unique task identifier
        repo:
          type: string
          description: Repository identifier
        title:
          type: string
          description: Task title
        description:
          type: string
          nullable: true
          description: Detailed task description
        estimate_points:
          type: integer
          minimum: 1
          maximum: 100
          description: Estimated effort in points
        priority:
          type: integer
          minimum: 0
          maximum: 1000
          description: Task priority (higher = more important)
        area:
          type: string
          maxLength: 120
          nullable: true
          description: Code area for conflict avoidance
        tier:
          type: integer
          minimum: 0
          maximum: 3
          default: 0
          description: Risk tier (0=safe, 3=high-risk)
        required_skills:
          type: array
          items:
            type: string
          description: Required skills for this task
        lease_expires_at:
          type: string
          format: date-time
          description: When the lease expires
    WorkResponse:
      type: object
      required: [worker_id, leases]
      properties:
        worker_id:
          type: string
          description: Worker identifier
        leases:
          type: array
          items:
            $ref: '#/components/schemas/LeaseView'
          description: Assigned task leases
    TaskStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [in_progress, blocked, pr_opened, merged]
          description: New task status
        message:
          type: string
          maxLength: 4000
          nullable: true
          description: Optional status message
        artifact:
          $ref: '#/components/schemas/TaskArtifact'

    # Admin Schemas
    RepoCreateRequest:
      type: object
      required: [repo]
      properties:
        repo:
          type: string
          minLength: 1
          maxLength: 200
          description: Repository key (e.g. 'demo' or 'owner/name')
        max_open_prs:
          type: integer
          minimum: 0
          maximum: 100
          default: 3
          description: Maximum open PRs (0 disables assignments)
        area_locks_enabled:
          type: boolean
          default: true
          description: Enable area-based conflict prevention
    RepoCreateResponse:
      type: object
      required: [ok, repo]
      properties:
        ok:
          type: boolean
          default: true
        repo:
          type: string
          description: Created repository identifier
    TaskCreateRequest:
      type: object
      required: [repo, title]
      properties:
        repo:
          type: string
          minLength: 1
          maxLength: 200
          description: Repository identifier
        title:
          type: string
          minLength: 1
          maxLength: 300
          description: Task title
        description:
          type: string
          nullable: true
          description: Detailed task description
        estimate_points:
          type: integer
          minimum: 1
          maximum: 100
          default: 1
          description: Estimated effort in points
        priority:
          type: integer
          minimum: 0
          maximum: 1000
          default: 10
          description: Task priority (higher = more important)
        required_skills:
          type: array
          items:
            type: string
          description: Required skills for this task
        area:
          type: string
          maxLength: 120
          nullable: true
          description: Code area for conflict avoidance
        tier:
          type: integer
          minimum: 0
          maximum: 3
          default: 0
          description: Risk tier (0=safe, 3=high-risk)
    TaskCreateResponse:
      type: object
      required: [ok, task_id]
      properties:
        ok:
          type: boolean
          default: true
        task_id:
          type: string
          description: Created task identifier

    # Advanced Feature Schemas
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
          additionalProperties: true
paths:
  /v1/workers/register:
    post:
      summary: Register a new worker
      description: Register a worker with skills and capacity to receive task assignments
      operationId: registerWorker
      tags: [Workers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWorkerRequest'
      responses:
        '201':
          description: Worker registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterWorkerResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/workers/heartbeat:
    post:
      summary: Send worker heartbeat
      description: Update worker status and maintain lease validity
      operationId: workerHeartbeat
      tags: [Workers]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '200':
          description: Heartbeat received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatResponse'
        '401':
          description: Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/work:
    get:
      summary: Fetch available work
      description: Get task assignments based on worker skills and capacity
      operationId: fetchWork
      tags: [Workers]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Available work assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkResponse'
        '401':
          description: Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tasks/{task_id}/status:
    post:
      summary: Update task status
      description: Update the status of an assigned task with optional artifacts
      operationId: updateTaskStatus
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: Task identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusUpdateRequest'
      responses:
        '200':
          description: Task status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
        '400':
          description: Invalid request data or task state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authorized to update this task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/repos:
    post:
      summary: Create or configure repository
      description: Register a repository with capacity limits and configuration
      operationId: createRepo
      tags: [Admin]
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoCreateRequest'
      responses:
        '201':
          description: Repository created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoCreateResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing admin token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/tasks:
    post:
      summary: Create a new task
      description: Create a task for worker assignment
      operationId: createTask
      tags: [Admin]
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreateResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing admin token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/state:
    get:
      summary: Get system state
      description: Retrieve current system status and metrics
      operationId: getSystemState
      tags: [Admin]
      security:
        - AdminToken: []
      responses:
        '200':
          description: System state retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  workers_online:
                    type: integer
                    description: Number of active workers
                  tasks_queued:
                    type: integer
                    description: Number of tasks waiting for assignment
                  tasks_in_progress:
                    type: integer
                    description: Number of tasks currently being worked on
                  repositories:
                    type: array
                    items:
                      type: object
                      properties:
                        repo:
                          type: string
                        max_open_prs:
                          type: integer
                        current_open_prs:
                          type: integer
        '401':
          description: Invalid or missing admin token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...